Extract external getrow to sdk
