+++ packages/server/src/api/controllers/row/external.ts
@@ -16,20 +16,6 @@ import {
 } from "@budibase/types"
 import sdk from "../../../sdk"
 
-async function getRow(
-  tableId: string,
-  rowId: string,
-  opts?: { relationships?: boolean }
-) {
-  const response = (await handleRequest(Operation.READ, tableId, {
-    id: breakRowIdField(rowId),
-    includeSqlRelationships: opts?.relationships
-      ? IncludeRelationship.INCLUDE
-      : IncludeRelationship.EXCLUDE,
-  })) as Row[]
-  return response ? response[0] : response
-}
-
 export async function handleRequest(
   operation: Operation,
   tableId: string,
@@ -71,7 +57,9 @@ export async function patch(ctx: UserCtx<PatchRowRequest, PatchRowResponse>) {
     id: breakRowIdField(id),
     row: rowData,
   })
-  const row = await getRow(tableId, id, { relationships: true })
+  const row = await sdk.rows.external.getRow(tableId, id, {
+    relationships: true,
+  })
   const table = await sdk.tables.getTable(tableId)
   return {
     ...response,
@@ -96,7 +84,9 @@ export async function save(ctx: UserCtx) {
   const responseRow = response as { row: Row }
   const rowId = responseRow.row._id
   if (rowId) {
-    const row = await getRow(tableId, rowId, { relationships: true })
+    const row = await sdk.rows.external.getRow(tableId, rowId, {
+      relationships: true,
+    })
     return {
       ...response,
       row,
@@ -109,7 +99,7 @@ export async function save(ctx: UserCtx) {
 export async function find(ctx: UserCtx) {
   const id = ctx.params.rowId
   const tableId = ctx.params.tableId
-  return getRow(tableId, id)
+  return sdk.rows.external.getRow(tableId, id)
 }
 
 export async function destroy(ctx: UserCtx) {
