+++ packages/server/src/sdk/app/rows/index.ts
@@ -2,10 +2,12 @@ import * as attachments from "./attachments"
 import * as rows from "./rows"
 import * as search from "./search"
 import * as utils from "./utils"
+import * as external from "./external"
 
 export default {
   ...attachments,
   ...rows,
   ...search,
-  utils: utils,
+  utils,
+  external,
 }
